feat(server): handling transpilation
