+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/constants/flat-field-metadata-relation-properties-to-compare.constant.ts
@@ -4,5 +4,5 @@ export const FLAT_FIELD_METADATA_RELATION_PROPERTIES_TO_COMPARE = [
   'label',
   'description',
   'isActive',
-  'standardOverrides'
+  'standardOverrides',
 ] as const satisfies FlatFieldMetadataPropertiesToCompare[];
